+++ chrome-extension/src/options/Options.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import './Options.css'
 
 function App() {
-  const [crx, setCrx] = useState('create-chrome-ext')
+  const [crx, setCrx] = useState('leakshield')
 
   return (
     <main>
@@ -10,7 +10,7 @@ function App() {
 
       <h6>v 0.0.0</h6>
 
-      <a href="https://www.npmjs.com/package/create-chrome-ext" target="_blank">
+      <a href="https://www.npmjs.com/package/leakshield" target="_blank">
         Power by {crx}
       </a>
     </main>