+++ chrome-extension/src/popup/Popup.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import './Popup.css'
 function App() {
-  const [crx, setCrx] = useState('create-chrome-ext')
+  const [crx, setCrx] = useState('leakshield')
 
   return (
     <main>